backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
groq==0.4.2
PyGithub==2.1.1
